+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 2.4.0 (TBD)
 
+## 2.3.1
+
+- Fixed a regression introduced in 2.3.0 causing PUT operations to encrypt files with wrong smkId
 - Added `workloadIdentityAzureClientId` config option allowing to customize Azure Client for `WORKLOAD_IDENTITY` authentication (snowflakedb/snowflake-connector-nodejs#1174)
 - Added `workloadIdentityImpersonationPath` config option for `authenticator=WORKLOAD_IDENTITY` allowing workloads to authenticate as a different identity through transitive service account impersonation (snowflakedb/snowflake-connector-nodejs#1178, snowflakedb/snowflake-connector-nodejs#1179, snowflakedb/snowflake-connector-nodejs#1182)